Fix rExp4j

+++ src/examples/rExp4j/ArrayStack.java
@@ -4,7 +4,7 @@ import java.util.EmptyStackException;
 
 class ArrayStack {
 
-    private double[] data;
+    private int[] data;
 
     private int idx;
 
@@ -18,13 +18,13 @@ class ArrayStack {
                     "Stack's capacity must be positive");
         }
 
-        data = new double[initialCapacity];
+        data = new int[initialCapacity];
         idx = -1;
     }
 
-    void push(double value) {
+    void push(int value) {
         if (idx + 1 == data.length) {
-            double[] temp = new double[(int) (data.length * 1.2) + 1];
+            int[] temp = new int[(int) (data.length * 1.2) + 1];
             System.arraycopy(data, 0, temp, 0, data.length);
             data = temp;
         }
@@ -32,14 +32,14 @@ class ArrayStack {
         data[++idx] = value;
     }
 
-    double peek() {
+    int peek() {
         if (idx == -1) {
             throw new EmptyStackException();
         }
         return data[idx];
     }
 
-    double pop() {
+    int pop() {
         if (idx == -1) {
             throw new EmptyStackException();
         }
